Properly checking entity being name in gae store. Making utils use FLASK_DATA_STORE env var before deciding on type of store to initialize

+++ tests/runtests.sh
@@ -1,9 +1,12 @@
 #!/bin/sh
 
-export DATASTORE_EMULATOR_HOST=localhost:8081
-kill -9 `ps -ef | grep datastore-emulator | grep -v grep | sed -e "s/  */:/g" | cut -d ':' -f 2 `
-rm -Rf ../dsemu
-gcloud beta emulators datastore start --data-dir=../dsemu &
+if [ "$1" = "reset" ]; then
+    echo "Resetting emulator..."
+    export DATASTORE_EMULATOR_HOST=localhost:8081
+    kill -9 `ps -ef | grep datastore-emulator | grep -v grep | sed -e "s/  */:/g" | cut -d ':' -f 2 `
+    rm -Rf ../dsemu
+    gcloud beta emulators datastore start --data-dir=../dsemu &
+fi
 
 echo "Running tests with mem store"
 export FLASK_DATA_STORE=mem
